requirements.txt
Django==5.1.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
python-dotenv==1.0.0
django-filter==23.3
Pillow==10.4.0